requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
click==8.1.3
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
tomli==2.0.1
